+++ src/fields/Categories.php
@@ -387,12 +387,7 @@ class Categories extends ElementField
                 'name' => 'showStructure',
                 'if' => 'displayType == "dropdown"',
             ]),
-            SchemaHelper::lightswitchField([
-                'label' => Craft::t('formie', 'Allow Multiple'),
-                'instructions' => Craft::t('formie', 'Whether this field should allow multiple options to be selected.'),
-                'name' => 'multi',
-                'if' => 'displayType == "dropdown"',
-            ]),
+            ...$this->defineElementFieldMultiSelectAppearanceSchema(),
             ...$this->defineElementFieldSearchableDropdownAppearanceSchema(),
             SchemaHelper::labelPosition($this),
             SchemaHelper::instructions(),